requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.3.4
async-generator==1.10
attrs==20.3.0
autopep8==1.5.6
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
commonmark==0.9.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.12
distlib==0.3.1
Django==3.2.3
django-extra-fields==3.0.2
django-filter==2.4.0
django-versatileimagefield==2.1
djangorestframework==3.12.4
docopt==0.6.2
drf-extra-fields==3.1.1
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
glob2==0.7
google-api-core==1.26.3
google-api-python-client==2.1.0
google-auth==1.28.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
gTTS==2.2.2
h5py==3.2.1
httplib2==0.19.1
humanize==3.5.0
idna==2.10
importlib-metadata==3.8.0
ipykernel==5.5.0
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
Js2Py==0.70
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keyboard==0.13.5
kiwisolver==1.3.1
mahotas==1.4.11
MarkupSafe==1.1.1
matplotlib==3.4.0
mistune==0.8.4
mpmath==1.2.1
mysql-connector-python==8.0.23
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
nose==1.3.7
notebook==6.3.0
numpy==1.20.1
oauthlib==3.1.0
opencv-python==4.5.1.48
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.2
pipwin==0.5.1
playsound==1.2.2
prometheus-client==0.9.0
prompt-toolkit==3.0.18
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycodestyle==2.7.0
pycparser==2.20
PyGithub==1.55
Pygments==2.8.1
pyjsparser==2.7.1
PyJWT==2.1.0
PyNaCl==1.4.0
pyparsing==2.4.7
PyPrind==2.11.2
pyrsistent==0.17.3
pySmartDL==1.3.4
python-dateutil==2.8.1
python-decouple==3.4
python-magic==0.4.22
python-magic-bin==0.4.14
python-speech-features==0.6
pytz==2021.1
pywin32==300
pywinpty==0.5.7
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
review==1510
reviews==0.1.8
rich==10.2.2
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.1
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
sqlparse==0.4.1
sympy==1.7.1
terminado==0.9.3
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.4
virtualenv==20.4.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.4.1
